+++ addOns/custompayloads/src/main/java/org/zaproxy/zap/extension/custompayloads/ExtensionCustomPayloads.java
@@ -54,6 +54,7 @@ public class ExtensionCustomPayloads extends ExtensionAdaptor {
         super.hook(extensionHook);
 
         extensionHook.addOptionsParamSet(getParam());
+        extensionHook.addApiImplementor(new CustomPayloadsApi(this));
 
         if (hasView()) {
             extensionHook.getHookView().addOptionPanel(getOptionsPanel());